Slight renaming and clean up.
