+++ test/Chirp.RazorTest/IntegrationTest.cs
@@ -1,0 +1,22 @@
+﻿namespace Chirp.RazorTest;
+
+public class IntegrationTest : IClassFixture<WebApplicationFactory<Program>>
+{
+    private readonly HttpClient _client;
+
+    public IntegrationTest(WebApplicationFactory<Program> factory)
+    {
+        Environment.SetEnvironmentVariable("RUNNING_TESTS", "true");
+        _client = factory.CreateClient();  // Create a test HTTP client for Chirp.Razor
+    }
+
+    [Fact]
+    public async Task GetHomePage_ReturnsSuccess()
+    {
+        // Act
+        var response = await _client.GetAsync("/");
+
+        // Assert
+        Assert.Equal(System.Net.HttpStatusCode.OK, response.StatusCode);
+    }
+}
